requirements.txt
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
futures==3.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
psycopg2==2.9.10
pyfiglet==1.0.2
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
setuptools==76.0.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
